chore(homepage): update icons and add namespace and app information for Jenkins, Firefly-iii, and Firefly-iii Importer
The icons for ChangeDetectionIO and Home Assistant have been updated to "changedetection-io.png" and "home-assistant.png" respectively.

For the Jenkins entry, the namespace has been added as "jenkins" and the app name has been set to "jenkins-server".

For the Firefly-iii entry, the namespace has been added as "firefly-iii" and the app name has been set to "firefly-iii".

For the Firefly-iii Importer entry, the namespace has been added as "firefly-iii" and the app name has been set to "importer".